# Tareas/practica3-rest/openapi.yaml
openapi: 3.0.3
info:
  title: API de Pedidos
  description: Servicio REST de gestión de pedidos e ítems (Parte 2)
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Servidor local v1

paths:
  /pedidos:
    get:
      summary: Lista pedidos con paginacion y filtros
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: limite
          in: query
          schema:
            type: integer
            default: 10
        - name: estado
          in: query
          schema:
            type: string
            enum: [PENDIENTE, PAGADO, ENVIADO, CANCELADO]
      responses:
        '200':
          description: Lista paginada de pedidos
        '400':
          description: Parámetros inválidos
        '500':
          description: Error interno

    post:
      summary: Crea un nuevo pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codigo, clienteId, total, estado, items]
              properties:
                codigo:
                  type: string
                  example: "PED-99999"
                clienteId:
                  type: string
                  example: "CLI-01"
                total:
                  type: number
                  example: 150.5
                estado:
                  type: string
                  example: "PENDIENTE"
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      producto:
                        type: string
                        example: "Item 1"
                      cantidad:
                        type: integer
                        example: 1
                      precio:
                        type: number
                        example: 150.5
      responses:
        '201':
          description: Pedido creado con éxito
          headers:
            Location:
              schema:
                type: string
              description: URI del nuevo pedido
        '400':
          description: Error de validación o JSON malformado
        '409':
          description: Conflicto por código duplicado
        '500':
          description: Error interno

  /pedidos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Obtiene un pedido por ID
      responses:
        '200':
          description: Pedido encontrado
        '404':
          description: Pedido no encontrado
    put:
      summary: Reemplaza completamente un pedido
      responses:
        '200':
          description: Pedido actualizado
        '400':
          description: Validación fallida
        '404':
          description: Pedido no encontrado
    delete:
      summary: Elimina un pedido
      responses:
        '204':
          description: Eliminado sin cuerpo
        '404':
          description: Pedido no encontrado